refacor(jina): 음악 추천 상세 페이지 page 디렉토리로 이동
