+++ src/entities/comment/model/type.ts
@@ -1,0 +1,6 @@
+export interface Comment {
+  userId: string;
+  rate: number; // 별점
+  content: string;
+  src?: string;
+}
