+++ src/entities/user/config/index.ts
@@ -1,0 +1,3 @@
+import defaultImage from 'shared/assets/image/default_profile_image.png';
+
+export const DEFAULT_PROFILE_IMAGE = defaultImage;
