+++ src/entities/user/ui/ProfileImage.tsx
@@ -1,13 +1,16 @@
 import styled from '@emotion/styled';
 
+import { DEFAULT_PROFILE_IMAGE } from 'entities/user/config';
+
 interface ProfileImageProps {
   width: number;
   height: number;
-  src: string;
+  alt?: string;
+  src?: string;
 }
 
 export const ProfileImage = ({ width, height, src }: ProfileImageProps) => {
-  return <Profile width={width} height={height} src={src} />;
+  return <Profile width={width} height={height} src={src || DEFAULT_PROFILE_IMAGE} />;
 };
 
 const Profile = styled.img<{ width: number; height: number }>`
