+++ src/features/musicRecommendationBoard/model/useTagInput.ts
@@ -1,6 +1,6 @@
 import { useMemo, useState } from 'react';
 
-import { updateCheckedSet } from 'features/musicRecommendation/lib/updateCheckedSet';
+import { updateCheckedSet } from 'features/musicRecommendationBoard/lib/updateCheckedSet';
 
 import { GENRE, Genre, GenreId } from 'entities/genre/model/genre';
 import { MOOD, Mood, MoodId } from 'entities/mood/model/mood';
