+++ src/pages/detail/config/anchor.ts
@@ -1,0 +1,2 @@
+export const ANCHOR_REVIEW = 'review'; // '리뷰 및 평점'으로 이동하는 버튼 href
+export const ANCHOR_CONTENT = 'content'; // '글 내용'으로 이동하는 버튼 href
