+++ src/pages/detail/config/filterOption.ts
@@ -1,0 +1,14 @@
+export const REVIEW_FILTER_OPTIONS = [
+  {
+    label: '제목',
+    value: 'title',
+  },
+  {
+    label: '별점순',
+    value: 'star',
+  },
+  {
+    label: '리뷰만 보기',
+    value: 'review',
+  },
+];
