+++ src/pages/detail/ui/MoreButton.tsx
