+++ src/pages/detail/ui/MusicInfo/LikeButton.tsx
