+++ src/pages/detail/ui/cover.png
