+++ src/pages/detail/ui/styled.ts
