--- src/shared/ui/Detail/anchor.ts
@@ -1,2 +1,0 @@
-export const review = 'review';
-export const content = 'content';
